requirements/requirements.txt
pillow~=11.3.0
tqdm~=4.67.1
numpy~=2.3.1
PyYAML~=6.0.2
scikit-learn~=1.7.1
matplotlib~=3.10.3
opencv-python~=4.12.0.88